Added basic confirm to other components where deleting is possible

+++ app/assets/javascripts/components/tours/SingleTour.es6.jsx
@@ -45,18 +45,20 @@ class SingleTour extends React.Component {
 
   handleDeleteDataFromServer(deleteUrl, e) {
     e.preventDefault();
-    $.ajax({
-      url: deleteUrl,
-      type: "DELETE",
-      dataType: "json",
-      success: function(data){
-        Materialize.toast(data, 3000, 'rounded');
-      }.bind(this),
-      error: function(err){
-        Materialize.toast('There was an issue deleting. Please contact admin.', 3000, 'rounded');
-        console.log(err);
-      }.bind(this)
-    });
+    if(confirm("Are you sure you wish to delete this record")) {
+      $.ajax({
+        url: deleteUrl,
+        type: "DELETE",
+        dataType: "json",
+        success: function(data){
+          Materialize.toast(data, 3000, 'rounded');
+        }.bind(this),
+        error: function(err){
+          Materialize.toast('There was an issue deleting. Please contact admin.', 3000, 'rounded');
+          console.log(err);
+        }.bind(this)
+      });
+    }
   }
 
   handlePostDataToServer(rankUrl, e) {
